+++ vasuki/unique/moment.py
@@ -1,14 +1,10 @@
 # -*- coding: utf-8 -*-
 # (c) 2019 Andreas Motl <andreas@terkin.org>
 # Apache License, Version 2
-"""
-``six-nibble-name`` converts six nibbles (three bytes) into a 4-character name.
-https://github.com/HowManyOliversAreThere/six-nibble-name
-"""
 import time
 
 from vasuki.format.sixnibblename.sixnibblename import integer_slug
 
 
-def generate_nibbleword():
+def generate_momentname():
     return integer_slug(time.time_ns())