+++ CHANGES.rst
@@ -12,10 +12,11 @@ in progress
 - Add Nagamani19 identifiers
 - Add Gibberish words
 - Add six-nibble-name integer slug
-- Add 4-character nibbleword generator
+- Add MomentName short epoch slug (4-character nibbleword generator)
 - Add release tooling
 - Expose generators via HTTP API
 - Improve performance of gibberish generator
+- Wording: Use ``MomentName`` for generator
 
 
 2019-06-28 0.0.0